pyproject.toml
[tool.poetry]
name = "ts-demo-pacman-backend"
version = "0.1.0"
description = ""
authors = ["dexsper"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.1.0"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
httpx = "^0.26.0"
coverage = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
